create interface for BoardView to call ChessGame funcs
Square now uses BoardToGameInterface class to set its mouse press events
in order to avoid creating weird cyclic references that would occur if
I gave BoardScene a reference to ChessGame.

I wanted to do this so that requests for possible moves could go through
ChessGame first, which could check whose turn it was. I did not want
the PieceMovements class to have to be aware of that, so I needed a way
to have ChessGame call PieceMovements functions in behalf of the view.

The following redesign allowed me to remove the reference to
PieceMovements from BoardScene, and give ChessGame more central control
of everything.