fix: prevent removeItem(None) from being called
On the first turn of the game, pieces are drawn on their initial
positions and Square.setPiecePixmap() is called a ton. This used to
call BoardScene.removeItem() to get rid of the previous pixmap if there
was a piece captured on that square. This would erroneously call
BoardScene.removeItem(None) on the first turn, so an additional check
was added to make sure it was not None.

Some innacurated docstrings were also removed.