expand on the Rook class
Maintains a list of all the squares it would be able to see were there
no pieces. As of right now it's called controlledSquares but that will
be refactored to something less misleading. Should be able to 'pin'
a piece, but that has no been tested, no doubt there are bugs.

All of initialization is done in the parent Piece class now.

When instantiated, a Piece will add itself to the passed in square,
which is how the instance will be saved. In the Piece __init__ func,
it is self.square.setPiece() that starts off the modification of the
board state by the pieces, because it calls their updateSquares()
method. This is because when a piece moves, Square.setPiece() will
invariably be called, so I figure that would be a good place to start
the modifications.