reimplement square highlighting when a piece is clicked
Slightly different to how it was before, but now the squares passed in
are just the names of the squares to be highlighted. How the view
determines it should be highlighting squares is also a little different,
as the model will return a JSON-like response so the view knows what to
in response to a click.