fix: update squares properly
When a piece moves, the squares have to update properly to reflect the
change, which is what I did here. There still might be some updates
missing.

This also properly recomputes the squares of pieces that may be affected
by the move.

Also fixed the rook's updateSquares() method, which forgot to clear
its moves and controlledSquares list.