introduce nonMoveControlledSquares and stop Kings from checking themselves
Non-move-controlled-squares are squares that a piece cannot move to
but still controls. This would be like a black queen defending a black
pawn. Before introducing this concept, the queen wouldn't have
controlled that square, which led to some bugs.

Adding this also allowed me to fully stop Kings from putting themselves
in check, which is nice.