fix double updates of pieces
Often, a piece would be update several times in one turn unnecessarily.
This was due to not clearing tracked squares before the move, which made
the piece show up in the oldSquareTrackers set. It was also due to
moving the piece before getting the trackers, which would also make the
piece show up in the newSquareTrackers.

Adding clearTrackedAndControlledSquares as well as deferring the move
to the end of the function fixed this.

It also fixed some weird behaviour with checks.