move piece before trackers are updated
In the previous commit, moving the piece was placed at the end of the
setSquare() function, which created an issue with certain pieces
thinking nothing was occupying that square and allowing illegal moves.

The piece is now moved after the trackers are obtained, but before the
trackers are updated, so as to avoid the issues in this commit as well
as the issues in the previous commit.