half complete adding castling
Did a lot of things. Created a Squares class for other classes to
easily get the squares.

Created a special_moves module that will handle castling and en passant.

Added some attributes to the king and rooks in order to add castling.
Also added a return to the King.setSquare() method so I can identify
when a king has castled.